+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 * Improved trill support (@rettinghaus)
+* Option for using encoded line breaks, but automatic page breaks (--breaks=line) (@earboxer)
 * Support for `trill@extender` (@rettinghaus)
 * Support for `note@head.visible` (@rettinghaus)
 * Improved spacing with crossing voices (@rettinghaus)